Falta hacer controles en varias rutas, y cambiar algunos endpoints.

+++ database/migrations/2020_05_12_054549_create_facturas_table.php
@@ -16,7 +16,7 @@ class CreateFacturasTable extends Migration
         Schema::create('facturas', function (Blueprint $table) {
             $table->id();
             $table->foreignId('pedido_id')->constrained()->onDelete('cascade');
-            $table->float('total');
+            $table->decimal('total',12);
         });
     }
